# docs/api/historial/openapi-1.2.0.yaml
...
openapi: 3.1.0
info:
  title: PideUTB API
  version: 1.2.0
  summary: Pedidos de comida dentro del campus universitario.
  description: |
    Superficie **síncrona** de PideUTB: consulta de menú, creación y consulta
    de pedidos, e inicio del cobro.

    La confirmación del pago **no ocurre aquí**: llega de forma asíncrona por
    webhook y está descrita en [`asyncapi.yaml`](asyncapi.yaml). El porqué de
    esa separación —y la alternativa descartada— está en
    [ADR-0003](../adr/0003-estrategia-integracion.md).

    ### Unidad monetaria

    Todos los importes son **enteros en centavos de peso colombiano**
    (`100` = 1 COP). No se usan decimales de punto flotante en ningún campo de
    dinero. El nombre de cada campo lleva el sufijo `_centavos` para que la
    unidad viaje en el contrato y no en la documentación: un campo llamado
    `precio` que a veces son pesos y a veces centavos es exactamente la clase
    de ambigüedad que este proyecto ya corrigió en su lenguaje ubicuo
    (arc42 §8.1).

    La elección de la unidad menor no es estética: es la que usa la pasarela de
    pagos, de modo que el número que guarda el pedido y el que se envía a
    cobrar son el mismo, sin conversión intermedia donde perder precisión.

    ### Lectura tolerante

    Un consumidor **debe ignorar los campos que no conozca**. Añadir un campo a
    una respuesta se considera un cambio compatible precisamente bajo esa
    condición; un consumidor que falle ante campos desconocidos rompe esa
...
servers:
...
paths:
...
  /v1/pedidos:
    post:
      operationId: crearPedido
      tags: [pedidos]
      summary: Crea un pedido.
      description: |
        El cliente solo decide **qué ítem** quiere y **cuántos**. El
        establecimiento se deriva del ítem y no se acepta del cliente: aceptarlo
        permitía asignar el pedido a un punto que no vende ese producto
        (`docs/violaciones.md`, V-01).

        Los campos no declarados en la petición **se ignoran** en lugar de
        producir un error. Esa decisión es deliberada y forma parte del
        contrato: un cliente que siga enviando `establecimiento_id` por
        costumbre no se rompe, pero tampoco consigue influir en el resultado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearPedidoRequest"
      responses:
        "201":
          description: Pedido creado, a la espera de pago.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pedido"
        "404":
          $ref: "#/components/responses/NoEncontrado"
        "409":
          $ref: "#/components/responses/Conflicto"
        "422":
          $ref: "#/components/responses/ErrorDeValidacion"

    get:
      operationId: listarPedidosDeEstablecimiento
      tags: [pedidos]
      summary: Lista los pedidos de un establecimiento.
      description: |
        Los pedidos que el mostrador tiene que atender, del más reciente al más
        antiguo: quien atiende necesita ver arriba lo que acaba de entrar, no lo
        que lleva media hora resuelto.

        Devuelve **todos** los estados, incluidos los ya entregados. Ocultarlos
        obligaría a recordar qué se acaba de entregar para poder corregirse, y
        [ESC-03](../arc42/arc42.md#esc-03) pide resolver en tres interacciones.

        `establecimiento_id` es **obligatorio** y no un filtro opcional: sin él,
        esta ruta devolvería los pedidos de todo el campus a quien preguntara.
      parameters:
        - name: establecimiento_id
          in: query
          required: true
          description: Establecimiento cuyos pedidos se consultan.
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        "200":
          description: Pedidos del establecimiento.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListaDePedidos"
        "422":
          $ref: "#/components/responses/ErrorDeValidacion"
  /v1/pedidos/{pedido_id}/estado:
    post:
      operationId: cambiarEstadoDePedido
      tags: [pedidos]
      summary: Avanza el estado de un pedido.
      description: |
        Mueve el pedido dentro de la máquina de estados del mostrador. Es la
        operación que cierra [ESC-03](../arc42/arc42.md#esc-03).

        **Transiciones permitidas:**

        | Desde | Hacia |
        |---|---|
        | `pagado` | `en_preparacion`, `cancelado` |
        | `en_preparacion` | `listo_para_recoger`, `cancelado` |
        | `listo_para_recoger` | `entregado` |

        Dos ausencias son deliberadas:

        - **`pendiente_pago` no es origen de ninguna.** El mostrador **no puede
          marcar un pedido como pagado**: esa transición solo la dispara el
          webhook firmado de la pasarela (`asyncapi.yaml`). Si esta operación
          pudiera hacerla, cualquiera comería gratis, porque todavía no hay
          autenticación (`docs/violaciones.md`, V-10).
        - **`entregado` y `cancelado` son finales.** Volver atrás desde ellos es
          un error de registro, no una operación del negocio, y permitirlo haría
          imposible contar cuántos pedidos se sirvieron.

        Es un sub-recurso con nombre y no un `PATCH` sobre el pedido: `PATCH`
        sugeriría que cualquier campo se puede modificar, y lo único que el
        mostrador puede tocar es el estado. La forma de la ruta dice qué está
        permitido.
      parameters:
        - $ref: "#/components/parameters/PedidoId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CambiarEstadoRequest"
      responses:
        "200":
          description: Estado actualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pedido"
        "404":
          $ref: "#/components/responses/NoEncontrado"
        "409":
          description: |
            La transición no está permitida desde el estado actual. Es `409` y
            no `422` porque la petición está bien formada: lo que no encaja es
            el estado del recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDeNegocio"
        "422":
          $ref: "#/components/responses/ErrorDeValidacion"
  /v1/pedidos/{pedido_id}:
    get:
      operationId: obtenerPedido
      tags: [pedidos]
      summary: Consulta un pedido y su estado.
      description: |
        Es el endpoint que el frontend consulta tras volver de la pasarela,
        porque la confirmación del pago llega por un canal distinto y puede no
        haber llegado todavía (ver [ADR-0003](../adr/0003-estrategia-integracion.md)
        y arc42 §6.3).
      parameters:
        - $ref: "#/components/parameters/PedidoId"
      responses:
        "200":
          description: El pedido existe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pedido"
        "404":
          $ref: "#/components/responses/NoEncontrado"
        "422":
          $ref: "#/components/responses/ErrorDeValidacion"
...
components:
  securitySchemes:
    FirmaPasarela:
      type: apiKey
      in: header
      name: X-Firma-Evento
      description: |
        Firma del cuerpo del evento calculada por la pasarela. Sin ella, el
        endpoint del webhook sería un punto público capaz de marcar pedidos como
        pagados.
  parameters:
...
    PedidoId:
      name: pedido_id
      in: path
      required: true
      description: Identificador del pedido.
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1
  responses:
    NoEncontrado:
      description: El recurso referenciado no existe.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDeNegocio"
    Conflicto:
      description: |
        La petición es válida pero el estado actual del sistema no permite
        atenderla (ítem no disponible, establecimiento cerrado, pedido ya
        pagado).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDeNegocio"
    ErrorDeValidacion:
      description: El cuerpo de la petición no cumple el esquema declarado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDeValidacion"
  schemas:
...
    CrearPedidoRequest:
      type: object
      required: [item_id]
      properties:
        item_id:
          type: integer
          format: int64
          minimum: 1
        cantidad:
          type: integer
          minimum: 1
          maximum: 50
          default: 1
          description: |
            El máximo existe para que un error de tecleo no genere un pedido
            imposible de preparar (`docs/violaciones.md`, V-03). Bajarlo en el
            futuro sería un cambio incompatible: dejaría de aceptarse una
            petición que hoy es válida.
      examples:
        - item_id: 1
          cantidad: 2
    EstadoPedido:
      type: string
      description: |
        Estados por los que pasa un pedido. Es un conjunto cerrado y no texto
        libre (`docs/violaciones.md`, V-08): mientras fuera texto libre, ningún
        consumidor podía saber qué valores debía saber tratar.

        **Quitar un valor de esta lista es un cambio incompatible.** Añadir uno
        también lo es en la práctica, porque un consumidor escrito contra la
        lista anterior no sabe qué hacer con el valor nuevo; por eso la política
        de versionado lo trata como incompatible salvo que se demuestre que
        ningún consumidor discrimina por estado.
      enum:
        - pendiente_pago
        - pagado
        - en_preparacion
        - listo_para_recoger
        - entregado
        - cancelado
    EstadoSolicitable:
      type: string
      description: |
        Estados que **el mostrador** puede pedir. Es un subconjunto de
        `EstadoPedido`, y está separado por dos motivos.

        **De contrato:** `EstadoPedido` viaja en las respuestas. Usarlo también
        en la petición lo dejaría en las dos direcciones, y quitarle un valor
        pasaría a ser incompatible por la regla **I-7** además de por la I-8,
        perdiendo la evolucionabilidad que la política le reconoce a un `enum`
        de respuesta.

        **De seguridad:** al no incluir `pendiente_pago` ni `pagado`, la regla
        de que el mostrador **no puede marcar un pedido como pagado** queda
        escrita aquí, legible sin ejecutar nada. Esa transición solo la dispara
        el webhook firmado de la pasarela.
      enum:
        - en_preparacion
        - listo_para_recoger
        - entregado
        - cancelado
    CambiarEstadoRequest:
      type: object
      description: |
        Estado al que el mostrador quiere llevar el pedido.

        Los dos modos de fallo quedan separados por código, y cada uno dice algo
        distinto: **`422`** significa «ese estado no es algo que puedas pedir»
        —es el caso de `pagado`—, y **`409`** significa «es pedible, pero no
        desde donde está el pedido», con el estado actual en el mensaje.
      required: [estado]
      properties:
        estado:
          $ref: "#/components/schemas/EstadoSolicitable"
      examples:
        - estado: en_preparacion
    ListaDePedidos:
      type: object
      description: |
        Pedidos de un establecimiento, del más reciente al más antiguo.

        Va envuelto en un objeto y no como array en la raíz, por el mismo motivo
        que `ListaDeItems`: con un sobre, añadir paginación o un total más
        adelante es un campo nuevo en una respuesta —cambio compatible—; con un
        array en la raíz sería una rotura.
      required: [establecimiento_id, pedidos]
      properties:
        establecimiento_id:
          type: integer
          minimum: 1
        pedidos:
          type: array
          items:
            $ref: "#/components/schemas/Pedido"
    Pedido:
      type: object
      required:
        - pedido_id
        - establecimiento_id
        - item_id
        - nombre_item
        - precio_unitario_centavos
        - cantidad
        - total_centavos
        - estado
        - codigo_canje
      properties:
        pedido_id:
          type: integer
          format: int64
          minimum: 1
        establecimiento_id:
          type: integer
          format: int64
          minimum: 1
        item_id:
          type: integer
          format: int64
          minimum: 1
        nombre_item:
          type: string
          description: |
            Instantánea del nombre en el momento de la compra. Si el
            establecimiento renombra el producto después, el pedido conserva lo
            que se compró (`docs/violaciones.md`, V-06).
        precio_unitario_centavos:
          type: integer
          minimum: 0
          description: Instantánea del precio en el momento de la compra.
        cantidad:
          type: integer
          minimum: 1
          maximum: 50
        total_centavos:
          type: integer
          minimum: 0
          description: |
            `precio_unitario_centavos × cantidad`. Se publica calculado en lugar
            de dejar que cada consumidor lo multiplique: es el importe que se
            cobra, y debe haber una sola fuente de verdad para él.
        estado:
          $ref: "#/components/schemas/EstadoPedido"
        codigo_canje:
          type: [string, "null"]
          pattern: "^[A-Z0-9]{6}$"
          description: |
            Código que el usuario presenta para recoger el pedido. Es `null`
            mientras el pedido no esté pagado.

            Se declara **siempre presente y anulable** en lugar de opcional: un
            consumidor encuentra la clave siempre y solo tiene que mirar si vale
            `null`. Un campo que a veces no viene obliga a distinguir «ausente»
            de «vacío», que es una fuente clásica de errores en el cliente.
      examples:
        - pedido_id: 1
          establecimiento_id: 1
          item_id: 1
          nombre_item: Arepa de huevo
          precio_unitario_centavos: 400000
          cantidad: 2
          total_centavos: 800000
          estado: pendiente_pago
          codigo_canje: null
...
    ErrorDeNegocio:
      type: object
      description: |
        Forma de error que emite el framework para los fallos de negocio. Se
        documenta tal como es y no como convendría que fuera: migrar a
        `application/problem+json` (RFC 9457) sería un cambio incompatible y
        requiere `/v2`.
      required: [detail]
      properties:
        detail:
          type: string
      examples:
        - detail: El ítem 999 no existe
    ErrorDeValidacion:
      type: object
      required: [detail]
      properties:
        detail:
          type: array
          items:
            type: object
            required: [loc, msg, type]
            properties:
              loc:
                type: array
                items:
                  type: [string, integer]
              msg:
                type: string
              type:
                type: string